Refactor background scripts for upcoming features
