+++ src/background/utils/tab-stats.js
