+++ src/store/options.js
@@ -11,15 +11,11 @@
 
 import { store } from 'hybrids';
 
-const UPDATE_ACTION_NAME = 'updateOptions';
-
-export function isUpdateOptionsMessage(message) {
-  return message?.action === UPDATE_ACTION_NAME;
-}
-
 export const DNR_RULES_LIST = chrome.runtime
   .getManifest()
   .declarative_net_request.rule_resources.map((r) => r.id);
+
+const UPDATE_OPTIONS_ACTION_NAME = 'updateOptions';
 
 const Options = {
   dnrRules: DNR_RULES_LIST.reduce(
@@ -28,6 +24,7 @@ const Options = {
   ),
   trackerWheelDisabled: false,
   wtmSerpReport: true,
+  version: '',
   [store.connect]: {
     async get() {
       const { options = {} } = await chrome.storage.local.get(['options']);
@@ -37,7 +34,10 @@ const Options = {
       await chrome.storage.local.set({ options });
 
       // Send update message to another contexts (background page / panel / options)
-      await chrome.runtime.sendMessage({ action: UPDATE_ACTION_NAME });
+      await chrome.runtime.sendMessage({
+        action: UPDATE_OPTIONS_ACTION_NAME,
+        options,
+      });
 
       return options;
     },
@@ -54,12 +54,8 @@ const Options = {
 
       DNR_RULES_LIST.forEach((rule) => {
         const enabled = options.dnrRules[rule];
-        if (enabled !== enabledRulesetIds.includes(rule)) {
-          if (enabled) {
-            enableRulesetIds.push(rule);
-          } else {
-            disableRulesetIds.push(rule);
-          }
+        if (enabledRulesetIds.includes(rule) !== enabled) {
+          (enabled ? enableRulesetIds : disableRulesetIds).push(rule);
         }
       });
 
@@ -75,8 +71,31 @@ const Options = {
 
 export default Options;
 
-chrome.runtime.onMessage.addListener((message) => {
-  if (isUpdateOptionsMessage(message)) {
+export async function observe(property, fn) {
+  let value;
+  const wrapperFn = (options) => {
+    if (value === undefined || options[property] !== value) {
+      value = options[property];
+      fn(value);
+    }
+  };
+
+  // Add listener for options updates
+  chrome.runtime.onMessage.addListener((msg) => {
+    if (msg.action === UPDATE_OPTIONS_ACTION_NAME) {
+      wrapperFn(msg.options);
+    }
+
+    return false;
+  });
+
+  // Run for the first time with the current options
+  const options = await store.resolve(store.get(Options));
+  wrapperFn(options);
+}
+
+chrome.runtime.onMessage.addListener((msg) => {
+  if (msg.action === UPDATE_OPTIONS_ACTION_NAME) {
     const options = store.get(Options);
 
     if (!store.pending(options)) {
